+++ packages/bridge/vue3-bridge/src/index.ts
@@ -1,3 +1,3 @@
 export { createBridgeComponent } from './provider';
-export { createRemoteComponent } from './create';
+export { createRemoteComponent, createRemoteAppComponent } from './create';
 export type { RenderFnParams } from '@module-federation/bridge-shared';