+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/rsbuild-plugin
 
+## 0.17.0
+
+### Patch Changes
+
+- Updated dependencies [3f736b6]
+  - @module-federation/node@2.7.9
+  - @module-federation/enhanced@0.17.0
+  - @module-federation/sdk@0.17.0
+
 ## 0.16.0
 
 ### Patch Changes