+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/utilities
 
+## 3.1.61
+
+### Patch Changes
+
+- @module-federation/sdk@0.17.0
+
 ## 3.1.60
 
 ### Patch Changes